LV1_PROBLEM_5_GALLERY_STORAGE_OPTIMIZER/requirements.txt
streamlit>=1.32.0
Pillow>=10.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
plotly>=5.18.0
pandas>=2.0.0
